requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
cachelib==0.9.0
celery
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.6
distro==1.9.0
faiss-cpu==1.8.0
Flask==3.0.3
Flask-Caching==2.3.0
frozenlist==1.4.1
GeoAlchemy2==0.15.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langsmith==0.1.63
MarkupSafe==2.1.5
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.30.3
orjson==3.10.3
packaging==23.2
psycopg2==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv
PyYAML==6.0.1
redis
regex==2024.5.15
requests==2.32.2
rq
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.0
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
